requirements.txt
fastapi==0.79.0
python-dotenv==0.20.0
sqlmodel==0.0.6
uvicorn==0.18.2
ukrdc-sqla==1.11.0
psycopg2==2.9.3
pytest==7.1.2